LoadXml in ajax service
