+++ addons/web/static/tests/component_extension_tests.js
@@ -3,7 +3,7 @@ odoo.define('web.component_extension_tests', function (require) {
 
     const makeTestEnvironment = require("web.test_env");
     const testUtils = require("web.test_utils");
-    const session = require('web.session');
+    const AjaxService = require('web.AjaxService');
 
     const { Component, tags } = owl;
     const { xml } = tags;
@@ -47,35 +47,31 @@ odoo.define('web.component_extension_tests', function (require) {
         });
 
         QUnit.test("Component lazy load template", async function (assert) {
-            assert.expect(7);
+            assert.expect(6);
             const fixture = document.body.querySelector('#qunit-fixture');
+            const ajaxService = new AjaxService();
+            ajaxService.globalProms = null;
+            ajaxService.loadFile = function (url) {
+                assert.step(url);
+                let template;
+                if (url === 'dependency1') {
+                    template = `<templates>
+                        <div t-name="Parent.test1" owl="1">
+                            <t t-call="Parent.test2"/>
+                        </div>
+                    </templates>`;
+                } else if (url === 'dependency2') {
+                    template = `<templates>
+                        <div t-name="Parent.test2" owl="1">
+                            lazy loaded 2
+                        </div>
+                    </templates>`;
+                }
+                return Promise.resolve(template);
+            }
             const env = {
-                session: {
-                    loadOwlXML: session.loadOwlXML,
-                    loadFile: function (url) {
-                        assert.step(url);
-                        let template;
-                        if (url === 'dependency1') {
-                            template = `<templates>
-                                <div t-name="Parent.test1" owl="1">
-                                    <t t-call="Parent.test2"/>
-                                </div>
-                            </templates>`;
-                        } else if (url === 'dependency2') {
-                            template = `<templates>
-                                <div t-name="Parent.test2" owl="1">
-                                    lazy loaded 2
-                                </div>
-                            </templates>`;
-                        } else if (url === 'dependency3') {
-                            template = `<templates>
-                                <div t-name="Parent.test3" owl="1">
-                                    Parent3
-                                </div>
-                            </templates>`;
-                        }
-                        return Promise.resolve(template);
-                    }
+                services: {
+                    ajax: ajaxService,
                 }
             }
             class Parent extends Component {}
@@ -99,15 +95,9 @@ odoo.define('web.component_extension_tests', function (require) {
             );
 
             parent.destroy();
-
-            class Parent2 extends Parent {}
-            Parent2.xmlDependencies = ['dependency3'].concat(Parent.xmlDependencies);
-
-            const parent2 = new Parent2();
+            const parent2 = new Parent();
             await parent2.mount(fixture);
-            assert.verifySteps([
-                'dependency3',
-            ])
+            assert.verifySteps([]);
 
             assert.strictEqual(
                 fixture.innerHTML,
